requirements.txt
pandas
requests
pyarrow
fastparquet
apache-airflow